CommitsBetween: learn to accept pagination params
When clients want the commits in a certain range, there's usually no
information available for them about how many commits are in this range.
The server will than parse and send commits as long as the client is
interested in them.

This change allows the client to signal how many commits at most the
server should send. It leverages the PaginationParameters structure as
previously exposed and used.

Requested as feature in: https://gitlab.com/gitlab-org/gitlab/-/issues/195667