Introduction of short-lived in-memory cache for reads distribution
With enabled distributed_reads feature each read operation leads to
a database query execution to get state of the storages for particular
repository. More reads leads to more database access, so pressure to
it increases linear (or worse).
To mitigate this pressure a short-living cache added before accessing
database. The expiration of the entries could be set via configuration
file: `distributed_reads_cache_expiration` key. The default is 1s.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3053