PostgreSQL notifications listener
Praefect should be able to connect directly to the PostgreSQL
or to the PgBouncer instance set in front of it. That is why
proxy_host and proxy_port configuration settings were added.
All connections need to be done via proxy if it is configured.
Exception to it is a listening connection that must be a direct
connection to the database instance.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3053