PostgreSQL notifications listener
As there is a need to follow for connection lifecycle events the simple
callback notification function was replaced with ListenHandler interface.
The Disconnect method of it will be triggered once connection to remote
would be lost.
Threshold of failing events now configurable with PostgresListenerOpts.
It now includes threshold for disconnect and connect attempt events.
As connections could be established and terminated (disconnect) and
we don't want an infinite loop of failed attempts to connect (connect
attempt).

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3053