ruby: Upgrade to Rugged 1.0
With v1.0, libgit2 released its first stable major version since
forever and as Rugged follows the same versioning scheme, it also
released v1.0 of the Gem. This caused us to be blocked on upgrading
libgit2/Rugged as some of our dependencies were depending on ~>0.28.

The two last dependencies which did depend on that are licensee and
gitlab-gollum-lib. This commit here thus upgrades both Gems, which
allows us to finally make use of rugged v1.0 and later.