Introduction of in-memory cache for reads distribution
With enabled distributed_reads feature each read operation leads to
a database query execution to get state of the storages for particular
repository. More read calls leads to more database access opearions,
so the pressure to it increases in linear (or even worse).
To mitigate this problem it was decided to introduce an in-memory cache
added before accessing the database. The expiration of the entries is
defaulted to 1min and invalidation happens on receiving events from the
database. The events are send by the triggers attached to the repositories
(delete operations) and storage_repositories (insert, delete, update) tables.

To monitor the cache a new counter was added: gitaly_praefect_uptodate_storages_cache_access_total.
It tracks amount of cache hits, misses and populates and evicts per virtual
repository. And to track an error rate of the notifications processing the
gitaly_praefect_uptodate_storages_errors_total was added with type set to one of:
sql_retrieve, notification_decode, cache_entry_removal.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3053