Shared connection used to listen for notifications
We require to listen for two notification channels from the
database for single purpose - invalidation of the cache with
up to date storages. A single connection gives a couple of
benefits compare to connection per channel approach: we
handle disconnection event properly and don't need to sync;
we have less direct connections to database.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3053